move hooks and clean for lint
