+++ src/app/[locale]/(logged-in)/organisation/profile/[tabId]/[subTabId]/components/SectorSiteAndWebsite/index.ts
@@ -1,0 +1,3 @@
+import SectorSiteAndWebsite from "./SectorSiteAndWebsite";
+
+export default SectorSiteAndWebsite;
