+++ src/app/[locale]/(logged-in)/organisation/profile/[tabId]/layout.tsx
@@ -1,17 +1,8 @@
 import { ReactNode } from "react";
 import { PageContainer } from "@/modules";
-import TabsContents from "./components/TabsContents";
 import TabsSections from "./components/TabsSections";
 
-export default function Layout({
-  children,
-  params,
-}: {
-  children: ReactNode;
-  params: { locale: string; tabId: string };
-}) {
-  const { tabId } = params;
-
+export default function Layout({ children }: { children: ReactNode }) {
   return (
     <PageContainer>
       <TabsSections />
